- cart list controller added and new api url added

+++ crafty_bay/lib/data/models/product_model.dart
@@ -10,7 +10,7 @@ class ProductModel {
   String? discountPrice;
   String? image;
   int? stock;
-  double? star;
+  int? star;
   String? remark;
   int? categoryId;
   int? brandId;
